Added Custom Exception, dto class and Validation

+++ src/main/java/com/employee/employee/service/EmployeeService.java
@@ -3,10 +3,14 @@ package com.employee.employee.service;
 import java.util.List;
 import java.util.Optional;
 
+import javax.validation.Validator;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.employee.employee.dto.EmployeeDto;
 import com.employee.employee.entity.Employee;
+import com.employee.employee.exception.EmployeeMgmtException;
 import com.employee.employee.repository.EmployeeRepository;
 
 
@@ -16,7 +20,20 @@ public class EmployeeService {
 	@Autowired
 	private EmployeeRepository empRepo;
 	
-	public Employee saveEmployee(Employee employee) {
+	@Autowired
+	private Validator validator;
+	
+	public Employee saveEmployee(EmployeeDto employeeDto) {
+		
+		validator.validate(employeeDto);
+		
+		Employee employee = new Employee();
+		
+		employee.setId(employeeDto.getId());
+		employee.setName(employeeDto.getName());
+		employee.setDepartment(employeeDto.getDepartment());
+		employee.setEmail(employeeDto.getEmail());
+		employee.setSalary(employeeDto.getSalary());
 		
 		Employee savedEmp = empRepo.save(employee);
 		
@@ -24,29 +41,42 @@ public class EmployeeService {
 	}
 	
 	
-	public List<Employee> getAllEmployees() {
+	public List<Employee> getAllEmployees() throws EmployeeMgmtException {
 		
 		List<Employee> empList = empRepo.findAll();
-	   
+		
+	    if(empList.isEmpty()) {
+	    	throw new EmployeeMgmtException("The Employee List is Empty", "4001");
+	    }
 		return empList;
 	}
 	
 	
-	public Employee getEmployeeById(Integer id) {
+	public Employee getEmployeeById(Integer id) throws EmployeeMgmtException {
 		
 		Optional<Employee> empOpt = empRepo.findById(id);
 		
 		if(empOpt.isEmpty()) {
-			throw new RuntimeException("Employee is not found");
+			
+			throw new EmployeeMgmtException("The given Employee ID is not present in the database", "4001");
+			
 		}
 		
 		return empOpt.get();
 		
 	}
 	
-	public String deleteEmployeeById(Integer id) {
+	public String deleteEmployeeById(Integer id) throws EmployeeMgmtException {
 		
-		empRepo.deleteById(id);
+		if(empRepo.existsById(id)) {
+			
+			empRepo.deleteById(id);
+			
+		}
+		else {
+			
+			throw new EmployeeMgmtException("The given Employee ID is not present in the database", "4001");
+		}
 		
 		return "Employee data deleted Successfully";
 	}
